Revert "[SPARK-46593][PS][TESTS] Refactor `data_type_ops` tests"
This reverts commit 3793c2f44099ccd10ed23e6a5d4c63734c788e6f.
