+++ python/pyspark/pandas/tests/connect/data_type_ops/test_parity_base.py
@@ -20,10 +20,7 @@ from pyspark.pandas.tests.data_type_ops.test_base import BaseTestsMixin
 from pyspark.testing.connectutils import ReusedConnectTestCase
 
 
-class BaseParityTests(
-    BaseTestsMixin,
-    ReusedConnectTestCase,
-):
+class BaseParityTests(BaseTestsMixin, ReusedConnectTestCase):
     pass
