+++ python/pyspark/pandas/tests/connect/data_type_ops/test_parity_binary_ops.py
@@ -17,14 +17,13 @@
 import unittest
 
 from pyspark.pandas.tests.data_type_ops.test_binary_ops import BinaryOpsTestsMixin
-from pyspark.pandas.tests.data_type_ops.testing_utils import OpsTestBase
+from pyspark.pandas.tests.connect.data_type_ops.testing_utils import OpsTestBase
+from pyspark.testing.pandasutils import PandasOnSparkTestUtils
 from pyspark.testing.connectutils import ReusedConnectTestCase
 
 
 class BinaryOpsParityTests(
-    BinaryOpsTestsMixin,
-    OpsTestBase,
-    ReusedConnectTestCase,
+    BinaryOpsTestsMixin, PandasOnSparkTestUtils, OpsTestBase, ReusedConnectTestCase
 ):
     pass
