+++ python/pyspark/pandas/tests/connect/data_type_ops/test_parity_null_ops.py
@@ -17,14 +17,13 @@
 import unittest
 
 from pyspark.pandas.tests.data_type_ops.test_null_ops import NullOpsTestsMixin
-from pyspark.pandas.tests.data_type_ops.testing_utils import OpsTestBase
+from pyspark.pandas.tests.connect.data_type_ops.testing_utils import OpsTestBase
+from pyspark.testing.pandasutils import PandasOnSparkTestUtils
 from pyspark.testing.connectutils import ReusedConnectTestCase
 
 
 class NullOpsParityTests(
-    NullOpsTestsMixin,
-    OpsTestBase,
-    ReusedConnectTestCase,
+    NullOpsTestsMixin, PandasOnSparkTestUtils, OpsTestBase, ReusedConnectTestCase
 ):
     pass
